+++ src/Data/Singletons/Promote/Type.hs
@@ -13,20 +13,21 @@ module Data.Singletons.Promote.Type
 import Language.Haskell.TH.Desugar
 import Data.Singletons.Names
 import Language.Haskell.TH
+import qualified Control.Monad.Fail as Fail
 
 -- the only monadic thing we do here is fail. This allows the function
 -- to be used from the Singletons module
-promoteType :: Monad m => DType -> m DKind
+promoteType :: Fail.MonadFail m => DType -> m DKind
 promoteType = go []
   where
-    go :: Monad m => [DTypeArg] -> DType -> m DKind
+    go :: Fail.MonadFail m => [DTypeArg] -> DType -> m DKind
     -- We don't need to worry about constraints: they are used to express
     -- static guarantees at runtime. But, because we don't need to do
     -- anything special to keep static guarantees at compile time, we don't
     -- need to promote them.
     go []       (DForallT _tvbs _cxt ty) = go [] ty
     go []       (DAppT (DAppT DArrowT (DForallT (_:_) _ _)) _) =
-      fail "Cannot promote types of rank above 1."
+      Fail.fail "Cannot promote types of rank above 1."
     go args     (DAppT t1 t2) = do
       k2 <- go [] t2
       go (DTANormal k2 : args) t1
@@ -56,15 +57,15 @@ promoteType = go []
       = return $ DConT tyFunArrowName `DAppT` k1 `DAppT` k2
     go _        ty@DLitT{} = pure ty
 
-    go args     hd = fail $ "Illegal Haskell construct encountered:\n" ++
-                            "headed by: " ++ show hd ++ "\n" ++
-                            "applied to: " ++ show args
+    go args     hd = Fail.fail $ "Illegal Haskell construct encountered:\n" ++
+                                 "headed by: " ++ show hd ++ "\n" ++
+                                 "applied to: " ++ show args
 
-promoteTypeArg :: Monad m => DTypeArg -> m DTypeArg
+promoteTypeArg :: Fail.MonadFail m => DTypeArg -> m DTypeArg
 promoteTypeArg (DTANormal t) = DTANormal <$> promoteType t
 promoteTypeArg ta@(DTyArg _) = pure ta -- Kinds are already promoted
 
-promoteUnraveled :: Monad m => DType -> m ([DKind], DKind)
+promoteUnraveled :: Fail.MonadFail m => DType -> m ([DKind], DKind)
 promoteUnraveled ty = do
   arg_kis <- mapM promoteType arg_tys
   res_ki  <- promoteType res_ty